[API] Added threadgroup size as a Opal_ComputePipelineDesc field
This allows setting threadgroup size per compute pipeline. This is necessary for Metal pipelines as there's no option to set threadgroup size in the shader like in other APIs.

On Vulkan / DirectX 12 / WebGPU threadgroup size must be set in the compute shader, so this field is ignored during pipeline creation although it can still be used for validation.